+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/info/display/contributor/field/FileEntryAuthorProfileImageInfoDisplayContributorField.java
@@ -14,13 +14,19 @@
 
 package com.liferay.document.library.web.internal.info.display.contributor.field;
 
+import com.liferay.info.display.contributor.field.BaseInfoDisplayContributorField;
 import com.liferay.info.display.contributor.field.InfoDisplayContributorField;
 import com.liferay.info.display.contributor.field.InfoDisplayContributorFieldType;
 import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
@@ -36,12 +42,12 @@ import org.osgi.service.component.annotations.Reference;
 	property = "model.class.name=com.liferay.portal.kernel.repository.model.FileEntry",
 	service = InfoDisplayContributorField.class
 )
-public class DLFileEntryAuthorNameInfoDisplayContributorField
-	implements InfoDisplayContributorField<FileEntry> {
+public class FileEntryAuthorProfileImageInfoDisplayContributorField
+	extends BaseInfoDisplayContributorField<FileEntry> {
 
 	@Override
 	public String getKey() {
-		return "authorName";
+		return "authorProfileImage";
 	}
 
 	@Override
@@ -49,24 +55,41 @@ public class DLFileEntryAuthorNameInfoDisplayContributorField
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			locale, getClass());
 
-		return LanguageUtil.get(resourceBundle, "author-name");
+		return LanguageUtil.get(resourceBundle, "author-profile-image");
 	}
 
 	@Override
 	public InfoDisplayContributorFieldType getType() {
-		return InfoDisplayContributorFieldType.TEXT;
+		return InfoDisplayContributorFieldType.IMAGE;
 	}
 
 	@Override
-	public String getValue(FileEntry fileEntry, Locale locale) {
+	public Object getValue(FileEntry fileEntry, Locale locale) {
 		User user = _userLocalService.fetchUser(fileEntry.getUserId());
 
-		if (user != null) {
-			return user.getFullName();
+		if (user == null) {
+			return StringPool.BLANK;
+		}
+
+		ThemeDisplay themeDisplay = getThemeDisplay();
+
+		if (themeDisplay != null) {
+			try {
+				return JSONUtil.put(
+					"url", user.getPortraitURL(getThemeDisplay()));
+			}
+			catch (PortalException pe) {
+				if (_log.isDebugEnabled()) {
+					_log.debug(pe, pe);
+				}
+			}
 		}
 
 		return StringPool.BLANK;
 	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		FileEntryAuthorProfileImageInfoDisplayContributorField.class);
 
 	@Reference
 	private UserLocalService _userLocalService;
