+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/info/display/contributor/field/FileEntryPreviewImageInfoDisplayContributorField.java
@@ -14,48 +14,57 @@
 
 package com.liferay.document.library.web.internal.info.display.contributor.field;
 
+import com.liferay.document.library.util.DLURLHelper;
 import com.liferay.info.display.contributor.field.InfoDisplayContributorField;
 import com.liferay.info.display.contributor.field.InfoDisplayContributorFieldType;
+import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
-import java.util.ResourceBundle;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Eudaldo Alonso
+ * @author Alejandro Tardín
  */
 @Component(
 	property = "model.class.name=com.liferay.portal.kernel.repository.model.FileEntry",
 	service = InfoDisplayContributorField.class
 )
-public class DLFileEntryDescriptionInfoDisplayContributorField
+public class FileEntryPreviewImageInfoDisplayContributorField
 	implements InfoDisplayContributorField<FileEntry> {
 
 	@Override
 	public String getKey() {
-		return "description";
+		return "previewImage";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
-			locale, getClass());
-
-		return LanguageUtil.get(resourceBundle, "description");
+		return LanguageUtil.get(
+			ResourceBundleUtil.getBundle(locale, getClass()), "preview-image");
 	}
 
 	@Override
 	public InfoDisplayContributorFieldType getType() {
-		return InfoDisplayContributorFieldType.TEXT;
+		return InfoDisplayContributorFieldType.IMAGE;
 	}
 
 	@Override
-	public String getValue(FileEntry fileEntry, Locale locale) {
-		return fileEntry.getDescription();
+	public Object getValue(FileEntry fileEntry, Locale locale) {
+		try {
+			return JSONUtil.put(
+				"url", _dlURLHelper.getImagePreviewURL(fileEntry, null));
+		}
+		catch (Exception pe) {
+			return null;
+		}
 	}
+
+	@Reference
+	private DLURLHelper _dlURLHelper;
 
 }
